requirements.txt
requests~=2.32.3
pandas~=2.2.3
openpyxl~=3.1.5
pyyaml~=6.0.2
yadisk~=3.3.0
python-dotenv~=1.0.1
pytest~=8.3.5
numpy~=2.3.0